Extract numpy SFT conversion helpers into open_instruct.numpy_dataset_conversion
Moves the HF-to-OLMo-core numpy mmap conversion logic out of
scripts/data/convert_sft_data_for_olmocore.py and into a new module
open_instruct/numpy_dataset_conversion.py so it can be imported by
downstream callers (e.g. the upcoming OLMo-core SFT main). The CLI
script keeps its argument surface and just delegates to the library.

Split out of #1620 (match-sft).
